Fix Some Bugs
- Чутка поправил настройку уровня, теперь удобнее закидывать конфиги (SceneInstaller в иерархии который)
- Изменил условия победы (когда кончаются ходы - переход на следующий уровень)

+++ Packages/com.singularitygroup.hotreload/Editor/CLI/LinuxCliController.cs
@@ -4,20 +4,20 @@ using System.IO;
 using System.Threading.Tasks;
 using Debug = UnityEngine.Debug;
 
-namespace SingularityGroup.HotReload.Editor.Cli {
-
-    class LinuxCliController : ICliController {
-        Process process;
+namespace SingularityGroup.HotReload.Editor.Cli
+{
+    internal class LinuxCliController : ICliController
+    {
+        private Process process;
 
         public string BinaryFileName => "CodePatcherCLI";
         public string PlatformName => "linux-x64";
         public bool CanOpenInBackground => true;
 
-        public Task Start(StartArgs args) {
+        public Task Start(StartArgs args)
+        {
             var startScript = Path.Combine(args.executableSourceDir, "hotreload-start-script.sh");
-            if (!File.Exists(startScript)) {
-                throw new FileNotFoundException(startScript);
-            }
+            if (!File.Exists(startScript)) throw new FileNotFoundException(startScript);
             File.WriteAllText(startScript, File.ReadAllText(startScript).Replace("\r\n", "\n"));
             CliUtils.Chmod(startScript);
 
@@ -26,8 +26,9 @@ namespace SingularityGroup.HotReload.Editor.Cli {
             title = title.Replace("'", "");
 
             var cliargsfile = Path.GetTempFileName();
-            File.WriteAllText(cliargsfile,args.cliArguments);
-            var codePatcherProc = Process.Start(new ProcessStartInfo {
+            File.WriteAllText(cliargsfile, args.cliArguments);
+            var codePatcherProc = Process.Start(new ProcessStartInfo
+            {
                 FileName = startScript,
                 Arguments =
                     $"--title \"{title}\""
@@ -41,32 +42,35 @@ namespace SingularityGroup.HotReload.Editor.Cli {
                 RedirectStandardOutput = true,
                 RedirectStandardError = true
             });
-            if (codePatcherProc == null) {
-                if (File.Exists(cliargsfile)) {
-                    File.Delete(cliargsfile);
-                }
+            if (codePatcherProc == null)
+            {
+                if (File.Exists(cliargsfile)) File.Delete(cliargsfile);
                 throw new Exception("Could not start code patcher process.");
             }
+
             codePatcherProc.BeginErrorReadLine();
             codePatcherProc.BeginOutputReadLine();
-            codePatcherProc.OutputDataReceived += (_, a) => {
-            };
+            codePatcherProc.OutputDataReceived += (_, a) => { };
             // error data can also mean we kill the proc beningly
-            codePatcherProc.ErrorDataReceived += (_, a) => {
-            };
+            codePatcherProc.ErrorDataReceived += (_, a) => { };
             process = codePatcherProc;
             return Task.CompletedTask;
         }
 
-        public async Task Stop() {
+        public async Task Stop()
+        {
             await RequestHelper.KillServer();
-            try {
+            try
+            {
                 // process.CloseMainWindow throws if proc already exited.
                 // also we just rely on the pid file it is fine
                 CliUtils.KillLastKnownHotReloadProcess();
-            } catch {
+            }
+            catch
+            {
                 //ignored
             }
+
             process = null;
         }
     }
